requirements.txt
MarkupSafe
Werkzeug
Jinja2
Flask
PyYAML
sqlalchemy
